Extend Pro-Gen importer functionality (rearranged & improved GUI to Notebook, clean up program (PyLint), added code due to remarks from Nick Hall).
