+++ gramps/plugins/lib/libplugins.gpr.py
@@ -41,7 +41,7 @@ fname = 'libcairodoc.py',
 authors = ["The Gramps project"],
 authors_email = ["http://gramps-project.org"],
 #load_on_reg = True
-  )
+)
 
 #------------------------------------------------------------------------
 #
@@ -58,7 +58,7 @@ status = STABLE,
 fname = 'libgedcom.py',
 authors = ["The Gramps project"],
 authors_email = ["http://gramps-project.org"],
-  )
+)
 
 #------------------------------------------------------------------------
 #
@@ -94,8 +94,9 @@ fname = 'libgrampsxml.py',
 authors = ["The Gramps project"],
 authors_email = ["http://gramps-project.org"],
 #load_on_reg = True
-  )
-  #------------------------------------------------------------------------
+)
+
+#------------------------------------------------------------------------
 #
 # libholiday
 #
@@ -111,7 +112,7 @@ fname = 'libholiday.py',
 authors = ["The Gramps project"],
 authors_email = ["http://gramps-project.org"],
 #load_on_reg = True
-  )
+)
 
 #------------------------------------------------------------------------
 #
@@ -129,7 +130,7 @@ fname = 'libhtmlbackend.py',
 authors = ["The Gramps project"],
 authors_email = ["http://gramps-project.org"],
 #load_on_reg = True
-  )
+)
 
 #------------------------------------------------------------------------
 #
@@ -165,7 +166,7 @@ fname = 'libhtml.py',
 authors = ["Gerald Britton"],
 authors_email = ["gerald.britton@gmail.com"],
 #load_on_reg = True
-  )
+)
 
 #------------------------------------------------------------------------
 #
@@ -237,6 +238,24 @@ authors_email = ["http://gramps-project.org"],
 
 #------------------------------------------------------------------------
 #
+# libprogen
+#
+#------------------------------------------------------------------------
+register(GENERAL,
+id    = 'libprogen',
+name  = "Pro-Gen lib",
+description =  _("Provides common functionality for Pro-Gen import"),
+version = '1.0',
+gramps_target_version = MODULE_VERSION,
+status = STABLE,
+fname = 'libprogen.py',
+authors = ["The Gramps project"],
+authors_email = ["http://gramps-project.org"],
+#load_on_reg = True
+)
+
+#------------------------------------------------------------------------
+#
 # libplaceview
 #
 #------------------------------------------------------------------------
@@ -268,6 +287,7 @@ fname = 'libsubstkeyword.py',
 authors = ["The Gramps project"],
 authors_email = ["http://gramps-project.org"],
 )
+
 #------------------------------------------------------------------------
 #
 # libtreebase
